# x.json2.decoder2: use json_str and json_len, intead of just json