--- a/vlib/x/json2/decoder2/tests/json_sumtype_test.v
+++ b/vlib/x/json2/decoder2/tests/json_sumtype_test.v
@@ -50,7 +50,7 @@ fn test_any_sum_type() {
 
 	assert json.decode[json2.Any]('1.1')! == json2.Any(f64(1.1))
 
-	// Uncomment this when #22693 is fixed
+	// Uncomment this when #22710 is fixed
 	// assert json.decode[[]json2.Any]('["1", "2", "3"]')! == [json2.Any('1'), json2.Any('2'), json2.Any('3')]
 	// assert json.decode[json2.Any]('["1", "2", "3"]')! == json2.Any([json2.Any('1'), json2.Any('2'),
 	// 	json2.Any('3')])