### Description
Allow to decode part of a string without need to create a substring

### Benchmark

```sh
V cache folder /home/hitalo/.vmodules/.cache was wiped.
V tmp.c and tests folder folder /tmp/v_1000 was wiped.
Starting benchmark...
max_iterations: 1000000

***Structure and maps***
 SPENT   645.359 ms in decoder2.decode[Stru](json_data)!
 SPENT  1204.135 ms in old_json.decode(Stru, json_data)!

 SPENT   730.546 ms in decoder2.decode[SumTypes](json_data)!
 SPENT  1161.810 ms in old_json.decode(SumTypes, json_data)!

 SPENT   127.497 ms in decoder2.decode[StructType[time.Time]](json_data_timestamp)!
 SPENT    83.384 ms in decoder2.decode[StructType[string]](json_data1)!
 SPENT    95.434 ms in old_json.decode(StructType[string], json_data1)!

 SPENT    82.203 ms in decoder2.decode[StructTypeOption[string]](json_data1)!
 SPENT    95.225 ms in old_json.decode(StructTypeOption[string], json_data1)!

 SPENT    50.005 ms in decoder2.decode[StructType[int]](json_data2)!
 SPENT    98.830 ms in old_json.decode(StructType[int], json_data2)!

 SPENT   194.058 ms in decoder2.decode[map[string]string](json_data1)!
 SPENT   192.357 ms in old_json.decode(map[string]string, json_data1)!


***arrays***
 SPENT   297.313 ms in decoder2.decode[[]int]('[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!
 SPENT   533.552 ms in old_json.decode([]int, '[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]')!


***simple types***
 SPENT    25.032 ms in decoder2.decode[int]('2')!
 SPENT    15.057 ms in decoder2.decode[bool]('true')!
 SPENT    94.516 ms in decoder2.decode[time.Time]('2022-03-11T13:54:25Z')!
 SPENT    74.930 ms in decoder2.decode[string]('"abcdefghijklimnopqrstuv"')!

***alias***
 SPENT    17.453 ms in decoder2.decode[IntAlias](2)!
 SPENT    73.759 ms in decoder2.decode[StringAlias]('"abcdefghijklimnopqrstuv"')!

***Sumtypes***
 SPENT    64.705 ms in decoder2.decode[SumTypes](2)!
 SPENT   131.476 ms in decoder2.decode[SumTypes]('"abcdefghijklimnopqrstuv"')!
 SPENT   645.622 ms in raw decode from HTTP request
 SPENT   736.365 ms in decode from HTTP request
```

### Valgrind
 `./../../../../../v wipe-cache && ./../../../../../v -prod bench.v -gc none -o b_out && valgrind -s ./b_out`
 
 Before
 ```sh
 ==4056878== 
==4056878== HEAP SUMMARY:
==4056878==     in use at exit: 42,134 bytes in 1,451 blocks
==4056878==   total heap usage: 2,279 allocs, 828 frees, 97,070 bytes allocated
==4056878== 
==4056878== LEAK SUMMARY:
==4056878==    definitely lost: 28,087 bytes in 798 blocks
==4056878==    indirectly lost: 14,028 bytes in 651 blocks
==4056878==      possibly lost: 0 bytes in 0 blocks
==4056878==    still reachable: 19 bytes in 2 blocks
==4056878==         suppressed: 0 bytes in 0 blocks
==4056878== Rerun with --leak-check=full to see details of leaked memory
==4056878== 
==4056878== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
 ```
 
 After
 ```sh
 ==4056018== 
==4056018== HEAP SUMMARY:
==4056018==     in use at exit: 24,334 bytes in 691 blocks
==4056018==   total heap usage: 2,249 allocs, 1,558 frees, 96,814 bytes allocated
==4056018== 
==4056018== LEAK SUMMARY:
==4056018==    definitely lost: 23,727 bytes in 598 blocks
==4056018==    indirectly lost: 588 bytes in 91 blocks
==4056018==      possibly lost: 0 bytes in 0 blocks
==4056018==    still reachable: 19 bytes in 2 blocks
==4056018==         suppressed: 0 bytes in 0 blocks
==4056018== Rerun with --leak-check=full to see details of leaked memory
==4056018== 
==4056018== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 0 from 0)
 ```